Style resource-click-fallthrough
With this change, clicking anywhere inside a resource tile once
attribution text is visible should lead to the resource's URL. Note that
the change works best for browsers that support pointer-events. For
other browsers, some parts of the resource tile may be clickable (where
the caption element does not extend), but mostly it should behave the
same.