Sticky header tweaks
This stops the header from popping into view for a moment when scrolling
down. The problem was that the CSS transition would immediately run
because transform(-100%) isn't the default state. Now we skip the
transition in some cases with an extra CSS class that is only applied
after the other ones.

It's weird but it works!